Checking a couple vars for truthiness.
I am using ImageMapster with angular. When the two variables
I have corrected inside the refreshSelections and clearHighlight
functions are not set, console errors are generated.

Exact situation is me calling:
$('#mapster').mapster('set', false, regionList.join());
where regionList.join() returns an empty string in one case. When
there *are* no previous selections, errors occur when the two vars
I corrected are attempted to be used while they are actually null.

I am guessing Mapster actually has safeguards against reaching this
code at a higher level when things are empty, but I'm using it only
as a simple highlight-storage area and tracking all state inside angular,
so probably that code is unable to kick in.